Bullet Collision
Exactly what it says on the tin, calculates bullet collision and now traces to enemies for logic. Still has bugs but shouldn't be too difficult. Also have the working of a Base class - Base in the context of a location/settlement, it's for player information such as lives, money, score, and the now refactored Heat system.