Algebra and Caches
Optimised Bitmap loading by offshunting the decoding of the Bitmaps into a Texture Cache. The cache dynamically stores new textures and decodes them on runtime start, upon doing so the textures are available for use via everything else. Additionally, there is an error texture to handle incorrect texture inputs.

Next on the list is both getting enemies to react to bullets, and refactor the texture system so that it is easier to follow and simpler to use. Refactoring will likely also influence other aspects of the project.